New items added, such as renamefuco, loungekey, and others.

+++ zodiac2.py
@@ -30,45 +30,34 @@ async def whatis(ctx):
     print(ctx.message.content[8:])
     if ctx.message.content[8:] == "Socialist" or ctx.message.content[8:] == "socialist" or ctx.message.content[8:] == "Socialists" or ctx.message.content[8:] == "socialists":
         await ctx.channel.send("Socialists believe in the public ownership of the means of production. They are commonly found on the far-left end of the compass, with varying degrees of authoritarianism/libertarianism.")
+    elif ctx.message.content[8:] == "Classical Liberal" or ctx.message.content[8:] == "classical liberal" or ctx.message.content[8:] == "classical lib" or ctx.message.content[8:] == "Classical liberal" or ctx.message.content[8:] == "Classical Lib" or ctx.message.content[8:] == "Classical lib":
+        await ctx.message.channel.send("Classical liberals adhere strongly to the rights of the individual, however, unlike libertarians, they are a lot more moderate on the free market. You'll usually find them straight down from center or near center.")
+    elif ctx.message.content[8:] == "Nationalist" or ctx.message.content[8:] == "nationalist":
+        await ctx.message.channel.send("Nationalists believe in putting their own country or group first above others. They are often protectionist, and come in many different variants from Civic Nationalists to NazBols to Ethnonationalists.")
+    elif ctx.message.content[8:] == "Right Libertarian" or ctx.message.content[8:] == "Libertarian" or ctx.message.content[8:] == "libertarian" or ctx.message.content[8:] == "right libertarian":
+        await ctx.message.channel.send("Libertarians tend to favor solutions that involve the free market and the individual over government intervention, detesting excess public spending and tax hikes. They are found everywhere on the bottom right of the compass.")
+    elif ctx.message.content[8:] == "Paleoconservative" or ctx.message.content[8:] == "paleocon" or ctx.message.content[8:] == "paleoconservative" or ctx.message.content[8:] == "paleocon":
+        await ctx.channel.send("Paleoconservatives, also known as the Old Right, date back to the early 20th century in their beliefs. They place social tradition and economic nonintervention above all else, often being in the top right of the compass.")
+    elif ctx.message.content[8:] == "LibtSoc" or ctx.message.content[8:] == "Libertarian Socialist" or ctx.message.content[8:] == "Libertarian Socialist" or ctx.message.content[8:] == "libertarian socialist":
+        await ctx.channel.send(" Libertarian Socialists/AnComs advocate for the abolition of state and private property as a means to liberate people.")
+    elif ctx.message.content[8:] == "SocDem" or ctx.message.content[8:] == "socdem" or ctx.message.content[8:] == "social democrat" or ctx.message.content[8:] == "Social Democrat" or ctx.message.content[8:] == "Social democrat" or ctx.message.content[8:] == "Socdem":
+        await ctx.channel.send("Social Democrats believe in using government as a tool to bring about social progress and change. They lean authoritative, intersect with social justice, and are typically found on the left.")
+    elif ctx.message.content[8:] == "green" or ctx.message.content[8:] == "Green" or ctx.message.content[8:] == "left libertarian":
+        await ctx.channel.send("Greens tend to be less about the economic liberty and more about the social and infrastructure liberty, especially on envrionmental issues. The open source community and net neutrality opponents are good examples of Greens.")
+    elif ctx.message.content[8:] == "Neoliberal" or ctx.message.content[8:] == "neolib" or ctx.message.content[8:] == "neoliberal":
+        await ctx.channel.send("Neoliberalism mixes aspects of progressivism and classical liberalism to create an ideology that's moderate, pragmatic, and often focused on values such as globalism, lower government spending, and deregulation.")
+    elif ctx.message.content[8:] == "centrist" or ctx.message.content[8:] == "Centrist":
+        await ctx.channel.send("Centrists don't particularly lean any noticeable amount on the left/right scale; their policies are often moderate, and they focus on compromise over ideological passion. They can be anywhere on the authoritarian/libertarian scale.")
+    elif ctx.message.content[8:] == "neocon" or ctx.message.content[8:] == "neoconservative" or ctx.message.content[8:] == "Neoconservative" or ctx.message.content[8:] == "Neocon":
+        await ctx.channel.send("Neoconservatism is a foreign policy theory that builds on the idea that it's a democratic country's duty to spread its values through foreign intervention. On most issues but foreign policy, they tend to be more moderate.")
+    elif ctx.message.content[8:] == "Paternalist" or ctx.message.content[8:]  == "paternalist":
+        await ctx.channel.send("Paternalists believe in the state taking a parent role, being authoritative on social policy, but interventionist economically.")
+    elif ctx.message.content[8:] == "Paleoliberal" or ctx.message.content[8:] == "Paleolib" or ctx.message.content[8:] == "paleoliberal" or ctx.message.content[8:] == "paleolib":
+        await ctx.channel.send("A paleoliberal or a RINO/Whig in the US is a term for a conservative who is a lot more moderate and centrist in their beliefs. They typically are found close to the center on issues of tradition, being economically to the right, typically being the parallel to Blue Dogs.")
+    elif ctx.message.content[8:] == "Theologue" or ctx.message.content[8:] == "theologue":
+        await ctx.channel.send("Theologues are socially conservative, but vary widely on their economic policies. This includes Distributists, Christian Democrats, and right-theologues. Their main deciding factor in voting is their religion.")
     else:
-        if ctx.message.content[8:] == "Classical Liberal" or ctx.message.content[8:] == "classical liberal" or ctx.message.content[8:] == "classical lib" or ctx.message.content[8:] == "Classical liberal" or ctx.message.content[8:] == "Classical Lib" or ctx.message.content[8:] == "Classical lib":
-            await ctx.message.channel.send("Classical liberals adhere strongly to the rights of the individual, however, unlike libertarians, they are a lot more moderate on the free market. You'll usually find them straight down from center or near center.")
-        else:
-            if ctx.message.content[8:] == "Nationalist" or ctx.message.content[8:] == "nationalist":
-                await ctx.message.channel.send("Nationalists believe in putting their own country or group first above others. They are often protectionist, and come in many different variants from Civic Nationalists to NazBols to Ethnonationalists.")
-            else:
-                if ctx.message.content[8:] == "Right Libertarian" or ctx.message.content[8:] == "Libertarian" or ctx.message.content[8:] == "libertarian" or ctx.message.content[8:] == "right libertarian":
-                    await ctx.message.channel.send("Libertarians tend to favor solutions that involve the free market and the individual over government intervention, detesting excess public spending and tax hikes. They are found everywhere on the bottom right of the compass.")
-                else:
-                    if ctx.message.content[8:] == "Paleoconservative" or ctx.message.content[8:] == "paleocon" or ctx.message.content[8:] == "paleoconservative" or ctx.message.content[8:] == "paleocon":
-                        await ctx.channel.send("Paleoconservatives, also known as the Old Right, date back to the early 20th century in their beliefs. They place social tradition and economic nonintervention above all else, often being in the top right of the compass.")
-                    else:
-                        if ctx.message.content[8:] == "LibtSoc" or ctx.message.content[8:] == "Libertarian Socialist" or ctx.message.content[8:] == "Libertarian Socialist" or ctx.message.content[8:] == "libertarian socialist":
-                            await ctx.channel.send(" Libertarian Socialists/AnComs advocate for the abolition of state and private property as a means to liberate people.")
-                        else:
-                            if ctx.message.content[8:] == "SocDem" or ctx.message.content[8:] == "socdem" or ctx.message.content[8:] == "social democrat" or ctx.message.content[8:] == "Social Democrat" or ctx.message.content[8:] == "Social democrat" or ctx.message.content[8:] == "Socdem":
-                                await ctx.channel.send("Social Democrats believe in using government as a tool to bring about social progress and change. They lean authoritative, intersect with social justice, and are typically found on the left.")
-                            else:
-                                if ctx.message.content[8:] == "green" or ctx.message.content[8:] == "Green" or ctx.message.content[8:] == "left libertarian":
-                                    await ctx.channel.send("Greens tend to be less about the economic liberty and more about the social and infrastructure liberty, especially on envrionmental issues. The open source community and net neutrality opponents are good examples of Greens.")
-                                else:
-                                    if ctx.message.content[8:] == "Neoliberal" or ctx.message.content[8:] == "neolib" or ctx.message.content[8:] == "neoliberal":
-                                            await ctx.channel.send("Neoliberalism mixes aspects of progressivism and classical liberalism to create an ideology that's moderate, pragmatic, and often focused on values such as globalism, lower government spending, and deregulation.")
-                                    else:
-                                        if ctx.message.content[8:] == "centrist" or ctx.message.content[8:] == "Centrist":
-                                            await ctx.channel.send("Centrists don't particularly lean any noticeable amount on the left/right scale; their policies are often moderate, and they focus on compromise over ideological passion. They can be anywhere on the authoritarian/libertarian scale.")
-                                        else:
-                                            if ctx.message.content[8:] == "neocon" or ctx.message.content[8:] == "neoconservative" or ctx.message.content[8:] == "Neoconservative" or ctx.message.content[8:] == "Neocon":
-                                                await ctx.channel.send("Neoconservatism is a foreign policy theory that builds on the idea that it's a democratic country's duty to spread its values through foreign intervention. On most issues but foreign policy, they tend to be more moderate.")
-                                            else:
-                                                if ctx.message.content[8:] == "Paternalist" or ctx.message.content[8:]  == "paternalist":
-                                                    await ctx.channel.send("Paternalists believe in the state taking a parent role, being authoritative on social policy, but interventionist economically.")
-                                                else:
-                                                    if ctx.message.content[8:] == "Paleoliberal" or ctx.message.content[8:] == "Paleolib" or ctx.message.content[8:] == "paleoliberal" or ctx.message.content[8:] == "paleolib":
-                                                        await ctx.channel.send("A paleoliberal or a RINO/Whig in the US is a term for a conservative who is a lot more moderate and centrist in their beliefs. They typically are found close to the center on issues of tradition, being economically to the right, typically being the parallel to Blue Dogs.")
-                                                    else:
-                                                        if ctx.message.content[8:] == "Theologue" or ctx.message.content[8:] == "theologue":
-                                                            await ctx.channel.send("Theologues are socially conservative, but vary widely on their economic policies. This includes Distributists, Christian Democrats, and right-theologues. Their main deciding factor in voting is their religion.")
+        await ctx.channel.send(ctx.message.content[8:] + " is not a role.")
 
 
 @z0diac.command()
@@ -154,18 +143,21 @@ async def give(ctx):
             currentvalue_giver = repkey.get(str(ctx.message.author.id))
             dollar_amount = ctx.message.content.find('>') + 1
             if currentvalue_giver - int(ctx.message.content[dollar_amount:]) >= 0:
-                repkey[str(ctx.message.author.id)] = (currentvalue_giver - int(ctx.message.content[dollar_amount:]))
-                f.seek(0)
-                json.dump(repkey, f)
-                f.truncate()
-                with open("dictionary.json", 'r+') as file:
-                    repkey2 = json.load(file)
-                    repkey2[str(ctx.message.raw_mentions[0])] = (currentvalue_recipient + int(ctx.message.content[dollar_amount:]))
-                    file.seek(0)
-                    json.dump(repkey2, file)
-                    file.truncate()
-                payment = ctx.message.content[(ctx.message.content.find('>') + 1):]
-                await ctx.channel.send('You have successfully gifted ' + (str(ctx.message.mentions[0]))[:-5] + payment + ' Effort Points.' )
+                if int(ctx.message.content[dollar_amount:]) > 0:
+                    repkey[str(ctx.message.author.id)] = (currentvalue_giver - int(ctx.message.content[dollar_amount:]))
+                    f.seek(0)
+                    json.dump(repkey, f)
+                    f.truncate()
+                    with open("dictionary.json", 'r+') as file:
+                        repkey2 = json.load(file)
+                        repkey2[str(ctx.message.raw_mentions[0])] = (currentvalue_recipient + int(ctx.message.content[dollar_amount:]))
+                        file.seek(0)
+                        json.dump(repkey2, file)
+                        file.truncate()
+                    payment = ctx.message.content[(ctx.message.content.find('>') + 1):]
+                    await ctx.channel.send('You have successfully gifted ' + (str(ctx.message.mentions[0]))[:-5] + payment + ' Effort Points.')
+                else:
+                    await ctx.channel.send('Gifts must be a positive number, nice try.')
             else:
                 print('You cannot afford that item!')
     except IndexError or ValueError:
@@ -186,9 +178,9 @@ async def buy(ctx):
                     json.dump(repkey, f)
                     f.truncate()
                     with open('bigdab.png','rb') as image:
-                        await ctx.channel.send(file=image2)
+                        await ctx.channel.send(file=image)
                 else:
-                    await ctx.channel.send('You need 5 EP to do that!')
+                    await ctx.channel.send('You need 25 EP to do that!')
             else:
                 await ctx.channel.send("You can't dab here!")
         elif ctx.message.content[5:15] == 'renamefuco':
@@ -196,7 +188,7 @@ async def buy(ctx):
                 fuco = ctx.guild.get_member(346821778829475861)
                 beforefuco = fuco.display_name
                 if ctx.channel != z0diac.get_channel(432574353822056448):
-                    dollar_amount = 5
+                    dollar_amount = 25
                     if currentvalue_customer - dollar_amount >= 0:
                         repkey[str(ctx.message.author.id)] = currentvalue_customer - dollar_amount
                         f.seek(0)
@@ -212,7 +204,40 @@ async def buy(ctx):
                         await ctx.channel.send("Too long.")
             else:
                 await ctx.channel.send("Nice try, FuCo.")
+        elif ctx.message.content[5:12] == 'subrole':
+            if ctx.channel != z0diac.get_channel(432574353822056448):
+                        dollar_amount = (75 + len(ctx.message.author.roles) * 50)
+                        if currentvalue_customer - dollar_amount >= 0:
+                            repkey[str(ctx.message.author.id)] = currentvalue_customer - dollar_amount
+                            f.seek(0)
+                            json.dump(repkey, f)
+                            f.truncate()
+                            await ctx.guild.create_role(name=ctx.message.content[12:])
+                            await ctx.message.author.add_roles()
+                            await ctx.channel.send('You have been given the' + ctx.message.content[12:] + ' role.')
+                        else:
+                            await ctx.channel.send('You need ' + str(dollar_amount) + ' EP to do that!')
+        elif ctx.message.content[5:7] == 'ad':
+            if ctx.channel != z0diac.get_channel(432574353822056448):
+                dollar_amount = 200
+                if currentvalue_customer - dollar_amount >= 0:
+                    repkey[str(ctx.message.author.id)] = currentvalue_customer - dollar_amount
+                    f.seek(0)
+                    json.dump(repkey, f)
+                    f.truncate()
+                    await z0diac.get_channel(447424864907689984).send("**By** **" + ctx.message.author.name + ":** " + ctx.message.content[8:])
+                else :
+                    await ctx.channel.send('You need 200 EP to do that!')
+        elif ctx.message.content[5:] == 'membership':
+            if ctx.channel != z0diac.get_channel(432574353822056448):
+                dollar_amount = 300
+                if currentvalue_customer - dollar_amount >= 0:
+                    await ctx.message.author.add_roles(discord.utils.get(ctx.guild.roles,id=447429561366478849))
+                    await ctx.channel.send('You now have access to lounge!')
+                else:
+                    await ctx.channel.send('You need 300 EP to do that!')
         else:
             await ctx.channel.send('Sorry, I did not get that.')
+
 
 z0diac.run("MzgzMzg1MTk4NTM0MDY2MTg3.DdOKNg.k92hJ_rHCjllSa_ehn81b2s9uNk")
